Change: Sepatrate TypeVarGen to other file
