+++ src/transform/type_infer/mod.rs
@@ -4,5 +4,6 @@ pub mod traits;
 pub mod type_;
 pub mod type_env;
 pub mod type_infer;
+pub mod type_var_gen;
 
 pub use self::type_infer::TypeInfer;
