+++ src/transform/type_infer/type_var_gen.rs
@@ -1,0 +1,30 @@
+//
+// this code is based on https://github.com/nwoeanhinnogaehr/algorithmw-rust
+//
+// Copyright 2016 Noah Weninger
+//
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+//
+
+use crate::ir::type_::{Type, TypeVarID};
+
+pub struct TypeVarGen {
+    supply: usize,
+}
+
+impl TypeVarGen {
+    pub fn new() -> TypeVarGen {
+        TypeVarGen { supply: 0 }
+    }
+    pub fn next(&mut self) -> TypeVarID {
+        let v = TypeVarID::with_usize(self.supply);
+        self.supply += 1;
+        v
+    }
+
+    pub fn new_variable(&mut self) -> Type {
+        Type::Variable(self.next())
+    }
+}
